Private Gists Created Using Web UI Turn Public
The possible values from the github gist API for public is false, true
or null.  json-read-from-string decodes these as true, json-false or
json-null.

We have been treating json-false as as truth value, which leads to this
mismatch.

This commit also fixes the issue where public gists created through this
package are setting the gist public attribute to null instead of false.

Fixes #45